backend/requirements.txt
fastapi==0.124.4
uvicorn[standard]==0.38.0
SQLAlchemy==2.0.36
pydantic[email]
psycopg[binary]
passlib[bcrypt] bcrypt==4.1.2
pydantic==2.12.5
PyJWT==2.9.0
python-multipart

websockets
wsproto